+++ nix/pkgs/haskell/materialized-unix/.plan.nix/plutus-core.nix
@@ -187,8 +187,9 @@
           "PlutusCore/Quote"
           "PlutusCore/MkPlc"
           "PlutusCore/Evaluation/Machine/Ck"
+          "PlutusCore/Evaluation/Machine/BuiltinCostModel"
+          "PlutusCore/Evaluation/Machine/CostModelInterface"
           "PlutusCore/Evaluation/Machine/ExBudget"
-          "PlutusCore/Evaluation/Machine/ExBudgeting"
           "PlutusCore/Evaluation/Machine/ExBudgetingDefaults"
           "PlutusCore/Evaluation/Machine/Exception"
           "PlutusCore/Evaluation/Machine/ExMemory"
