+++ plutus-core/plutus-core.cabal
@@ -43,8 +43,9 @@ library
         PlutusCore.Quote
         PlutusCore.MkPlc
         PlutusCore.Evaluation.Machine.Ck
+        PlutusCore.Evaluation.Machine.BuiltinCostModel
+        PlutusCore.Evaluation.Machine.CostModelInterface
         PlutusCore.Evaluation.Machine.ExBudget
-        PlutusCore.Evaluation.Machine.ExBudgeting
         PlutusCore.Evaluation.Machine.ExBudgetingDefaults
         PlutusCore.Evaluation.Machine.Exception
         PlutusCore.Evaluation.Machine.ExMemory
