+++ plutus-core/plutus-core/src/PlutusCore/Evaluation/Machine/MachineParameters.hs
@@ -6,20 +6,20 @@ where
 
 import           PlutusCore.Constant
 
-import           PlutusCore.Core.Type                      hiding (Type)
-import           PlutusCore.Evaluation.Machine.ExBudget    ()
-import           PlutusCore.Evaluation.Machine.ExBudgeting
+import           PlutusCore.Core.Type                           hiding (Type)
+import           PlutusCore.Evaluation.Machine.BuiltinCostModel
+import           PlutusCore.Evaluation.Machine.ExBudget         ()
 
-import           GHC.Types                                 (Type)
+import           GHC.Types                                      (Type)
 
 {-| We need to account for the costs of evaluator steps and also built-in function
    evaluation.  The models for these have different structures and are used in
-   different parts of the code, so most of the time we pass separate objects
-   about.  It's convenient for clients of the evaluator to only have to worry
-   about a single object, so the CostModel type bundles the two together.  We
-   could conceivably have different evaluators with different internal costs, so
-   we keep the machine costs abstract.  The model for Cek machine steps is in
-   UntypedPlutusCore.Evaluation.Machine.Cek.CekMachineCosts.
+   different parts of the code, so inside the valuator we pass separate objects
+   about most of the time .  It's convenient for clients of the evaluator to
+   only have to worry about a single object, so the CostModel type bundles the
+   two together.  We could conceivably have different evaluators with different
+   internal costs, so we keep the machine costs abstract.  The model for Cek
+   machine steps is in UntypedPlutusCore.Evaluation.Machine.Cek.CekMachineCosts.
 -}
 data CostModel machinecosts =
     CostModel {
